Fix issue where snapshot was written to wrong file
At some point we switched from directly using the manifest to using
a copy of the manifest stored in the .frepo file, this introduced a
snapshot bug where the snapshot is stored to the local copy rather
than to the repository.

This patch fixes that issue by storing the snapshot in the manifest
directory.